requirements.txt

soundfile
tqdm

numpy>=1.16.4
PyYAML>=5.0
pandas>=0.23.4
pb_bss_eval>=0.0.2

numba==0.48.0
librosa==0.7.2

torch==1.8.0
torchaudio==0.8.0
pytorch-lightning==1.0.1
asteroid==0.5.1